.claude/worktrees/trusting-matsumoto-b85b61/requirements.txt
# requirements.txt
Flask[async]==3.0.3
markdown-it-py==3.0.0
Pygments==2.18.0
playwright==1.44.0
gunicorn==22.0.0  # WSGI/ASGI server (Dockerfile CMD); not imported
uvicorn==0.30.1  # provides uvicorn.workers.UvicornWorker for gunicorn (Dockerfile CMD); not imported
asgiref==3.8.1
unstructured[all-docs]==0.14.5
pdfminer.six==20221105
PyMuPDF==1.24.1
deepgram-sdk==5.1.0
google-cloud-texttospeech==2.21.0
google-genai>=1.0.0
requests==2.31.0
pydub==0.25.1
redis==5.0.1
rq==1.16.0
pdfplumber==0.10.4
camelot-py==0.11.0
img2table==1.4.2
opencv-python-headless==4.10.0.84  # transitive dep of camelot-py + img2table; pinned to lock the cv2 wheel ABI
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
nh3==0.2.18
# Test deps (run `pytest` locally)
pytest>=8.0
pytest-asyncio>=0.23
responses>=0.25